pyproject.toml
[project]
name = "derivden"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "hvplot>=0.12.2",
    "marimo>=0.20.4",
    "perspective-python>=4.3.0",
    "polars>=1.38.1",
    "pyarrow>=23.0.1",
    "ruff>=0.15.5",
    "websocket-client>=1.9.0",
]
